Added internal ingress + a test agent VM
This introduces a test agent VM in a separate network,
peers that network with the kubernetes network,
and allows the agent to talk to the internal LB IP.

There are lots more ports open and more traffic
allowed than necessary.

The agent VM is blank / idle. The deployment flow hasn't
been finalized, this is just for evaluating feasibility of
the method.