Always pull controller image for GKE
With the previous 'IfNotPresent' setting, the deploy step doesn't check
whether or not the image+tag that is currently being used is the latest
version - if there was an image+tag present within the cluster and that
image+tag had been updated in the repo, then redeploying the
application would not pull the new image.

Switching to 'Always' policy avoids those problems.