Created account for GCE plugin
It is not automatically added to Jenkins.
It needs to be added as a "Google Service Account
from private key", and the Kubernetes Credentials Provider
plugin doesn't know how to create keys of that type.
Instead, the user must manually create a private key
using GCP tools and then add the key to Jenkins
with an appropriate name.