Point plastic client.conf's WorkspaceServer to server
Both controller and agents run with the Plastic `cm` client present,
but no local server installed. Without a local server, some `cm`
operations will attempt to talk to the server indicated by
WorkspaceServer. One example is when doing `cm workspace create`
on Linux; the workspace creation will succeed, but the command will
error out with "Connection refused" as it attempts to connect to
the workspace server just afterward. This results in first checkout to
a fresh agent always failing on dynamic + plastic + Linux jobs.

We avoid the above errors by pointing the WorkspaceServer to
the cloud org.