Install + start controller on VM via Ansible, WIP
TF creates an Ansible account with an SSH keypair
Ansible can be used to log in and install software
onto the machine, and start with docker-compose

We need to figure out which user & uid/gid that
the jenkins controller should run as.

Also, we need to figure out how we should run the
container; just docker-compose up? or have a
surrounding systemd service that we control via systemctl?

Finally, we haven't solved injection of anything like
GCP credentials, Plastic config, or CasC material.