Change to ExecStopPost, to ensure container cleanup
A failed docker compose (because health check fails to go good) fails
the command, but leaves containers running. This means that the service
is in a non-running state, but containers are running.

We move the stop logic to ExecStopPost; this step is run both when
stopping a service, but also when handling the failed start of a service.